optimize schema and db session calls

+++ sync-service/api/models.py
@@ -3,34 +3,34 @@ import datetime
 
 
 class Title(db.Model):
-    id = db.Column(db.BigInteger, primary_key=True)
+    id = db.Column(db.Integer, primary_key=True)
     string = db.Column(db.String(200))
 
 
 class Subtitle(db.Model):
-    id = db.Column(db.BigInteger, primary_key=True)
+    id = db.Column(db.Integer, primary_key=True)
     string = db.Column(db.String(300))
 
 
 class FullYear(db.Model):
-    id = db.Column(db.BigInteger, primary_key=True)
+    id = db.Column(db.Integer, primary_key=True)
     string = db.Column(db.String(48))
 
 
 class Composer(db.Model):
-    id = db.Column(db.BigInteger, primary_key=True)
+    id = db.Column(db.Integer, primary_key=True)
     string = db.Column(db.String(32))
     first_name = db.Column(db.String(32))
     last_name = db.Column(db.String(32))
 
 
 class Piece(db.Model):
-    id = db.Column(db.BigInteger, primary_key=True, autoincrement=True)
-    title = db.Column(db.BigInteger)
-    subtitle = db.Column(db.BigInteger)
+    id = db.Column(db.Integer, primary_key=True, autoincrement=True)
+    title = db.Column(db.Integer)
+    subtitle = db.Column(db.Integer)
     year = db.Column(db.Integer)
-    full_year = db.Column(db.BigInteger)
-    composer = db.Column(db.BigInteger)
+    full_year = db.Column(db.Integer)
+    composer = db.Column(db.Integer)
     duration = db.Column(db.Integer)
